requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
sqlalchemy==2.0.23
sqlite3
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
faiss-cpu==1.7.4
boto3==1.34.0
